# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ouvrier-Express",
        "description": "Application de gestion de relation prestataire-client",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ajouterPresta": {
            "post": {
                "tags": [
                    "Profil prestataire"
                ],
                "summary": "Ajouter profil",
                "operationId": "641afae848b35a700928fcc2bdda057d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "metier": {
                                        "type": "string"
                                    },
                                    "disponibilite": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Création reussie"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/affichPresta{id}": {
            "get": {
                "tags": [
                    "Profil prestataire"
                ],
                "summary": "Voir profil",
                "operationId": "970971d10f5533d7a6115e615895811e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Afficher un prestataire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/api/modifPresta/{prestataire}": {
            "patch": {
                "tags": [
                    "Profil prestataire"
                ],
                "summary": "Modificier un profil",
                "operationId": "fe93a99ec056c9ea78fa60693a8c3e4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Modifier à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "metier": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil modifiée avec succés"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supprimPresta/{id}": {
            "patch": {
                "tags": [
                    "Profil prestataire"
                ],
                "summary": "Supprimer un profil",
                "operationId": "6a757c1e828c33ee01d1405d62172708",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Supprimer un profil à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            }
        },
...
        "/api/ajoutPrestaService": {
            "post": {
                "tags": [
                    "Prestation de service"
                ],
                "summary": "Ajouter une prestation de service",
                "operationId": "2f1ce5f884b8f665be38a6a9c501b092",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nomService": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prestation de service ajoutée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/affichPrestaService/{id}": {
            "get": {
                "tags": [
                    "Prestation de service"
                ],
                "summary": "Voir une prestation de service",
                "operationId": "834dda159bd38dac925275229b9e9860",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Prestation de service à afficher à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/modifPrestaService/{prestatationservice}": {
            "patch": {
                "tags": [
                    "Prestation de service"
                ],
                "summary": "Modificier une prestation de service",
                "operationId": "28f088c0ff81615ad170d6eef087542e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la prestation à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nomService": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prestation de service modifiée avec succés"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supprimPrestaService/{id}": {
            "patch": {
                "tags": [
                    "Prestation de service"
                ],
                "summary": "Supprimer une prestation de service",
                "operationId": "0352b6cbf98b1d038969fbd1c26281ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Suppression d'une prestation de service à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}